feat: wire up vello renderer

+++ macros/src/lib.rs
@@ -8,7 +8,7 @@ pub fn derive_view_base(input: TokenStream) -> TokenStream {
     let name = input.ident;
     let expanded = quote! {
         impl #name {
-            pub fn size(mut self, width: i32, height: i32) -> Self {
+            pub fn size(mut self, width: f64, height: f64) -> Self {
                 self.view_base.size.width = width;
                 self.view_base.size.height = height;
                 self
@@ -19,33 +19,33 @@ pub fn derive_view_base(input: TokenStream) -> TokenStream {
                 self
             }
 
-            pub fn padding_top(mut self, amount: i32) -> Self {
+            pub fn padding_top(mut self, amount: f64) -> Self {
                 self.view_base.padding_top = amount;
                 self
             }
 
-            pub fn padding_bottom(mut self, amount: i32) -> Self {
+            pub fn padding_bottom(mut self, amount: f64) -> Self {
                 self.view_base.padding_bottom = amount;
                 self
             }
 
-            pub fn padding_left(mut self, amount: i32) -> Self {
+            pub fn padding_left(mut self, amount: f64) -> Self {
                 self.view_base.padding_left = amount;
                 self
             }
 
-            pub fn padding_right(mut self, amount: i32) -> Self {
+            pub fn padding_right(mut self, amount: f64) -> Self {
                 self.view_base.padding_right = amount;
                 self
             }
 
-            pub fn padding_horizontal(mut self, padding: i32) -> Self {
+            pub fn padding_horizontal(mut self, padding: f64) -> Self {
                 self.view_base.padding_left = padding;
                 self.view_base.padding_right = padding;
                 self
             }
 
-            pub fn padding_vertical(mut self, padding: i32) -> Self {
+            pub fn padding_vertical(mut self, padding: f64) -> Self {
                 self.view_base.padding_top = padding;
                 self.view_base.padding_bottom = padding;
                 self
@@ -57,11 +57,11 @@ pub fn derive_view_base(input: TokenStream) -> TokenStream {
                 self.view_base.size
             }
 
-            fn width(&self) -> i32 {
+            fn width(&self) -> f64 {
                 self.view_base.size.width
             }
 
-            fn height(&self) -> i32 {
+            fn height(&self) -> f64 {
                 self.view_base.size.height
             }
 
@@ -69,27 +69,27 @@ pub fn derive_view_base(input: TokenStream) -> TokenStream {
                 self.view_base.visible
             }
 
-            fn padding_top(&self) -> i32 {
+            fn padding_top(&self) -> f64 {
                 self.view_base.padding_top
             }
 
-            fn padding_bottom(&self) -> i32 {
+            fn padding_bottom(&self) -> f64 {
                 self.view_base.padding_bottom
             }
 
-            fn padding_left(&self) -> i32 {
+            fn padding_left(&self) -> f64 {
                 self.view_base.padding_left
             }
 
-            fn padding_right(&self) -> i32 {
+            fn padding_right(&self) -> f64 {
                 self.view_base.padding_right
             }
 
-            fn padding_horizontal(&self) -> i32 {
+            fn padding_horizontal(&self) -> f64 {
                 self.view_base.padding_left + self.view_base.padding_right
             }
 
-            fn padding_vertical(&self) -> i32 {
+            fn padding_vertical(&self) -> f64 {
                 self.view_base.padding_top + self.view_base.padding_bottom
             }
         }
